Moved the following functions into training.utils - get_pretrained_model_details - get_training_param_from_session_state - get_segmentation_model_funcs - get_segmentation_model_func2params
